Unimplemented framework for Nugget backed HALs
These HALs will communicate with NuggetOS. Each HAL operates with a
differnet app meaning they can reside in separate daemons without
needing to synchronize with each other.

They are currently unimplemented but the framework is in place to build
them and have them running if added to the device config.

Change-Id: Id42fb0eda3bbf7b0d1d243d98cfe40dac5a76a9f