Plumbing for Citadel connection in HALs.
The HALs are developed against the Nugget interface which could be
implemented in a number of ways. Currently, we have Citadel as the only
implementation but libnugget allows for abstraction of that detail.

The HAL services are the point at which the implementation is defined and
injected into the HAL implementation. This also allows the HALs to be unit
tested with a mocked connection.

CitadelClient needs to be implemented (probably using the not-yet-done
libcitadel) to complete the connection.

Change-Id: I42234fa9a6fb6b392b03c4e78543375ded062495