keymaster: implementation and test for ImportKey
This is an initial implementation for ImportKey,
some support libraries, and a test.

The implementation includes some input validation
code that would need to be duplicated in citadel, and
thus could be removed from here (but will leave for
now, and update depending on how the other methods
shape up).

Change-Id: I1cd554d6101fd60b740b7665e3d9889138c79c5b