Citadeld: handle requests on multiple threads
Ensure that each app only has one client at a time but allow different
apps to be accessed concurrently. The driver is thread safe so there is
no need to synchronize driver access.

This also allows a hard reset request to be processed if there is a
blocking transaction going on.

Bug: 77728719
Change-Id: I75d5ce8d23d7562d7b6e718ff4925ffad0b637a3