NUGGET_PARAM_ENABLE_UPDATE tells us whether it worked or not
Giving the wrong password still fails, but now with the correct
password this call now returns a boolean indicating whether the
header was CHANGED to valid. If it was, the AP should request or
force a Citadel reboot to give the new image a chance to run. If
nothing changed, there's no need to do that.

Bug: b/72983372
Test: Manual update, entered PIN, repeat

Change-Id: I763d9519c4f5cfa93ac513f13baf5d53f0b3bae6
Signed-off-by: Bill Richardson <wfrichar@google.com>