keymaster: always pass update() calls through
Pass update() calls even when there is no input data.
This is done so as to handle AES-GCM, where there may
be no input-data, but AAD data in the associated params.

As an optimization the HAL may choose to inspect the
algorithm and block-mode and do the needful, but this
is not currently done.

Bug: 109771020
Test: EncryptionOperationsTest.AesGcmMultiPartAad passes
Change-Id: I53a37cb65324eda2e2576f308d748fa9e77961f0