keymaster: abort operations on HAL errors
Not all error code paths in the HAL are
aborting initiated operations.  This change
tackles the error returns in attest().

This particular error case has not been encountered
in the field, but any errors during attest would
result in a similar situtation as described in the
associated bug -- operation handle leakage on Citadel.

Bug: 116055338
Test: release-tests.sh pass on PVT1
Change-Id: Icae96cc502cf16ece29c87e84c7b54c3e3e43727