Register citadeld with the power.stats service
This allows the power.stats service to retrieve citadeld's cached
stats information from time to time.

Both citadeld and power.stats are services, but AIDL prefers to
have services talk to clients and vice-versa.

This CL adds a new StatsDelegate class to provide the client-like
interface that power.stats requires, and instantiates it in a
background thread so that citadeld won't block if the power.stats
service is not available.

Bug: 70510004
Test: manual

These commands are handy for testing all the possible
combinations of events:

  logcat | egrep -i 'citadel|power'

  /vendor/bin/hw/citadel_updater -lv
  lshal debug android.hardware.power.stats@1.0::IPowerStats/default

  killall android.hardware.power.stats@1.0-service.pixel
  killall citadeld

  setprop ctl.stop vendor.power.stats-hal-1-0
  setprop ctl.start vendor.power.stats-hal-1-0

  setprop ctl.stop vendor.citadeld
  setprop ctl.start vendor.citadeld

Change-Id: I207ea4773faf9856bc7e39da8b4168002f23174e
Signed-off-by: Bill Richardson <wfrichar@google.com>